+++ react/features/base/connection/index.js
@@ -4,3 +4,4 @@ export * from './actions';
 export * from './actionTypes';
 export * from './constants';
 export * from './functions';
+export * from './utils';
