+++ react/features/invite/components/dial-in-info-page/DialInInfoApp.web.js
@@ -5,12 +5,16 @@ import { I18nextProvider } from 'react-i18next';
 import { isMobileBrowser } from '../../../base/environment/utils';
 import { i18next } from '../../../base/i18n';
 import { parseURLParams } from '../../../base/util/parseURLParams';
+import { DIAL_IN_INFO_PAGE_PATH_NAME } from '../../constants';
 import { DialInSummary } from '../dial-in-summary';
 
 import NoRoomError from './NoRoomError';
 
 document.addEventListener('DOMContentLoaded', () => {
     const { room } = parseURLParams(window.location, true, 'search');
+    const { href } = window.location;
+    const ix = href.indexOf(DIAL_IN_INFO_PAGE_PATH_NAME);
+    const url = (ix > 0 ? href.substring(0, ix) : href) + room;
 
     ReactDOM.render(
         <I18nextProvider i18n = { i18next }>
@@ -18,7 +22,8 @@ document.addEventListener('DOMContentLoaded', () => {
                 ? <DialInSummary
                     className = 'dial-in-page'
                     clickableNumbers = { isMobileBrowser() }
-                    room = { decodeURIComponent(room) } />
+                    room = { decodeURIComponent(room) }
+                    url = { url } />
                 : <NoRoomError className = 'dial-in-page' /> }
         </I18nextProvider>,
         document.getElementById('react')
