+++ react/features/invite/components/dial-in-summary/web/DialInSummary.js
@@ -3,7 +3,7 @@
 import React, { Component } from 'react';
 
 import { translate } from '../../../../base/i18n';
-import { doGetJSON } from '../../../../base/util';
+import { getDialInConferenceID, getDialInNumbers } from '../../../_utils';
 
 import ConferenceID from './ConferenceID';
 import NumbersList from './NumbersList';
@@ -29,6 +29,11 @@ type Props = {
      * The name of the conference to show a conferenceID for.
      */
     room: string,
+
+    /**
+     * The url where we were loaded.
+     */
+    url: string,
 
     /**
      * Invoked to obtain translated strings.
@@ -177,7 +182,7 @@ class DialInSummary extends Component<Props, State> {
             return Promise.resolve();
         }
 
-        return doGetJSON(`${dialInConfCodeUrl}?conference=${room}@${mucURL}`, true)
+        return getDialInConferenceID(dialInConfCodeUrl, room, mucURL, new URL(this.props.url))
             .catch(() => Promise.reject(this.props.t('info.genericError')));
     }
 
@@ -191,20 +196,12 @@ class DialInSummary extends Component<Props, State> {
         const { room } = this.props;
         const { dialInNumbersUrl, hosts } = config;
         const mucURL = hosts && hosts.muc;
-        let URLSuffix = '';
 
         if (!dialInNumbersUrl) {
             return Promise.reject(this.props.t('info.dialInNotSupported'));
         }
 
-        // when room and mucURL are available
-        // provide conference when looking up dial in numbers
-
-        if (room && mucURL) {
-            URLSuffix = `?conference=${room}@${mucURL}`;
-        }
-
-        return doGetJSON(`${dialInNumbersUrl}${URLSuffix}`, true)
+        return getDialInNumbers(dialInNumbersUrl, room, mucURL)
             .catch(() => Promise.reject(this.props.t('info.genericError')));
     }
